+++ src/index.ts
@@ -1,2 +1,4 @@
+export {Options as QROptions, ErrorCorrectionLevel} from '@bloomprotocol/qr'
+
 export * from './types'
 export {renderRequestElement} from './renderRequestElement'