+++ src/renderRequestElement.ts
@@ -1,36 +1,33 @@
 import Bowser from 'bowser'
+import {Options as QROptions} from '@bloomprotocol/qr'
 
-import {QROptions, ButtonOptions, RequestData, ShouldRenderButton, RequestElementResult} from './types'
+import {ButtonOptions, RequestData, ShouldRenderButton, RequestElementResult} from './types'
 import {renderRequestButton} from './elements/renderRequestButton'
 import {renderRequestQRCode} from './elements/renderRequestQRCode'
 
-export const renderRequestElement = (config: {
+export const renderRequestElement = (_config: {
   container: HTMLElement
   requestData: RequestData
   qrOptions?: Partial<QROptions>
   shouldRenderButton?: ShouldRenderButton
   buttonOptions: ButtonOptions
 }): RequestElementResult => {
-  if (typeof config.shouldRenderButton === 'undefined') {
-    config.shouldRenderButton = parsedResult => {
-      const isSupportedPlatform = parsedResult.platform.type === 'mobile' || parsedResult.platform.type === 'tablet'
-      const isSupportedOS = parsedResult.os.name === 'iOS' || parsedResult.os.name === 'Android'
+  const {shouldRenderButton, ...config} = _config
 
-      return isSupportedPlatform && isSupportedOS
-    }
-  }
+  let renderButton: boolean
 
-  let shouldRenderButton: boolean
+  if (typeof shouldRenderButton === 'undefined') {
+    const parsedResult = Bowser.parse(window.navigator.userAgent)
 
-  if (typeof config.shouldRenderButton === 'boolean') {
-    shouldRenderButton = config.shouldRenderButton
+    const isSupportedPlatform = parsedResult.platform.type === 'mobile' || parsedResult.platform.type === 'tablet'
+    const isSupportedOS = parsedResult.os.name === 'iOS' || parsedResult.os.name === 'Android'
+
+    renderButton = isSupportedPlatform && isSupportedOS
+  } else if (typeof shouldRenderButton === 'boolean') {
+    renderButton = shouldRenderButton
   } else {
-    shouldRenderButton = config.shouldRenderButton(Bowser.parse(window.navigator.userAgent))
+    renderButton = shouldRenderButton(Bowser.parse(window.navigator.userAgent))
   }
 
-  if (shouldRenderButton) {
-    return renderRequestButton(config)
-  } else {
-    return renderRequestQRCode(config)
-  }
+  return (renderButton ? renderRequestButton : renderRequestQRCode)(config)
 }